refactor to separate ParentLoanMapper and ChildLoanMapper for single responsibility

+++ src/main/java/org/matthiaskarl/techassessment/creditservice/domain/Product.java
@@ -4,7 +4,7 @@ public record Product(
         long id,
         String name,
         String type,
-        double amount,
+        long amount,
         String currencyCode,
         double interestRate,
         String startDate,
