+++ src/main/java/org/matthiaskarl/techassessment/creditservice/domain/RealSecurity.java
@@ -3,7 +3,7 @@ package org.matthiaskarl.techassessment.creditservice.domain;
 public record RealSecurity(
         String type,
         String address,
-        double collateralValue,
+        long collateralValue,
         String currency,
         String nextRevaluationDate
 ) {
